requirements.txt
# Global requirements for pn-metrics project
PyYAML>=6.0
requests>=2.31.0
